safelife/requirements.txt
tensorflow-gpu==1.13.1
matplotlib==3.0.3
scikit-image==0.15.0
pyemd==0.5.1
numpy>=1.11.0
scipy>=1.0.0
gym>=0.12.5
imageio>=2.5.0
ipython>=5.0
pyglet==1.3.2
